Fixed the Unit Testing
