+++ test/main.test.ts
@@ -1,33 +1,21 @@
 import '@aws-cdk/assert/jest';
-import { App, AppProps } from '@aws-cdk/core';
+import { App } from '@aws-cdk/core';
 import { MagentoStack } from '../src/main';
 
-
-test('Snapshot', () => {
-
-const appProps: AppProps = {
-  analyticsReporting: false,
-  autoSynth: false,
-  context: {
-    contextKey: {
+test('For Mandatory Infra Constructs have been created', () => {
+  const app = new App({
+    context: {
       route53_domain_zone: 'magento.mydomain.com',
       magento_debug_task: 'yes',
     },
-  },
-  outdir: 'outdir',
-  runtimeInfo: false,
-  stackTraces: false,
-  treeMetadata: false,
-};
-
-  const app = new App(appProps);
+  });
 
   const stackName = process.env.CDK_STACK_NAME ? process.env.CDK_STACK_NAME : 'magento';
   const clusterName = stackName;
 
   const devEnv = {
-    account: process.env.CDK_DEFAULT_ACCOUNT,
-    region: process.env.CDK_DEFAULT_REGION,
+    account: '1234567890',
+    region: 'us-east-1',
   };
 
   const stack = new MagentoStack(app, stackName, {
@@ -35,7 +23,14 @@ const appProps: AppProps = {
     createCluster: true,
     env: devEnv,
   });
-
-  expect(stack).not.toHaveResource('AWS::S3::Bucket');
+  // Check for Mandatory Resources
+  expect(stack).toHaveResource('AWS::RDS::DBInstance');
+  expect(stack).toHaveResource('AWS::RDS::DBCluster');
+  expect(stack).toHaveResource('AWS::OpenSearchService::Domain');
+  expect(stack).toHaveResource('AWS::ECS::Service');
+  expect(stack).toHaveResource('AWS::ElasticLoadBalancingV2::LoadBalancer');
+  
+  // Expect for Resource with this Specs 
+  
   expect(app.synth().getStackArtifact(stack.artifactId).template).toMatchSnapshot();
 });
